# By Sebastian Hoehna (54) and others
# Via Michael Landis (6) and others
* 'development' of https://github.com/revbayes/revbayes: (92 commits)
  keep dnCDBDP from throwing error when not setting psi/phi
  improvements/fixes to graph flip edge, shift edge moves
  Writing Nexus files with standard states using spaces in state labels (not sure anymore if that caused a bug)
  Adding the option to add vectors to a string (mostly for printing purposes)
  give user the option to specify whether the simulated character matrix should have gaps/missing characters matching the ones in the clamped matrix (the default) or not (then no gaps/missing in the simulated matrix)
  parse ambiguous natural number states
  add new graph edge updates
  allow Proposal to inform MH move which elements of nodes are actually touched -- something like touch specialization, but Sebastian might hate this :-)
  initial work for random graph distributions/moves
  ch-seq model speed-up
  modify AdjacentRateModifier to work with Natural[][] due to some tricky type conversion problems for Real[][] -> Natural[][]
  add VectorBinarySwitch to update sets of Bernoulli variables
  small edit to getAffectedSites
  speed updates to tree history classes
  Removing whitespaces from sequences
  Adding a PIC function when having multiple samples per species
  small edit to allow for a single tree in ancestral state mapping instead of a tree trace
  Adding a function to compute independent contrasts. This is used for posterior predictive testing.
  fixes to likelihood/simulate for GeneralTreeHistoryCtmc
  Some work on Codon models
  ...

+++ src/core/analysis/mcmc/output/Trace.h
@@ -65,7 +65,7 @@ namespace RevBayesCore {
         
     protected:
         
-        int                             burnin;
+        size_t                          burnin;
         std::string                     fileName;
         std::string                     parmName;
         std::vector<valueType>          values;                                     //!< the values of this trace
