+++ src/revlanguage/analysis/mcmc/RlMcmcmc.cpp
@@ -59,11 +59,11 @@ void Mcmcmc::constructInternalObject( void )
         mvs.push_back( ws_vec_mvs[i].getValue() );
     }
     const std::string &                                     sched   = static_cast<const RlString &>( moveschedule->getRevObject() ).getValue();
-    int                                                     nchains = static_cast<const Natural &>( num_chains->getRevObject() ).getValue();
-    int                                                     si      = static_cast<const Natural &>( swap_interval->getRevObject() ).getValue();
+    int                                                     nchains = (int)static_cast<const Natural &>( num_chains->getRevObject() ).getValue();
+    int                                                     si      = (int)static_cast<const Natural &>( swap_interval->getRevObject() ).getValue();
     double                                                  delta   = static_cast<const RealPos &>( delta_heat->getRevObject() ).getValue();
-    int                                                     nreps   = static_cast<const Natural &>( num_runs->getRevObject() ).getValue();
-    int                                                     ntries  = static_cast<const Natural &>( num_init_attempts->getRevObject() ).getValue();
+    int                                                     nreps   = (int)static_cast<const Natural &>( num_runs->getRevObject() ).getValue();
+    int                                                     ntries  = (int)static_cast<const Natural &>( num_init_attempts->getRevObject() ).getValue();
 
     RevBayesCore::Mcmcmc *m = new RevBayesCore::Mcmcmc(mdl, mvs, mntr, sched, nchains, si, delta, ntries);
