+++ src/revlanguage/functions/DistributionFunctionRv.h
@@ -156,7 +156,7 @@ RevLanguage::RevPtr<RevVariable> RevLanguage::DistributionFunctionRv<valueType>:
         
     }
     
-    int n = static_cast<const Natural &>( this->args[0].getVariable()->getRevObject() ).getValue();
+    int n = (int)static_cast<const Natural &>( this->args[0].getVariable()->getRevObject() ).getValue();
     ModelVector<valueType> *values = new ModelVector<valueType>();
     for (int i = 0;  i < n; ++i)
     {
